create useCurrentUser hook
